pyproject.toml
[project]
name = "lumagen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cartesia>=1.0.13",
    "numpy>=2.1.1",
    "openai>=1.51.0",
    "pyairtable>=2.3.3",
    "python-dotenv>=1.0.1",
    "rich>=13.9.1",
    "soundfile>=0.12.1",
]


[project.scripts]
main = "lumagen.main:main"
clear = "lumagen.main:clear_state"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
